refactor(proxy): split meeting and registrant client methods
Move active meeting and registrant ITX API calls into dedicated files
matching the domain ITXMeetingClient and ITXRegistrantClient interfaces.
